created test_model stage and added dvc documentation
