+++ .gitignore
@@ -1,4 +1,5 @@
 /data
 /models
+/reports
 /venv
 /env
